await stop bugfixes

+++ aoget/web/file_queue.py
@@ -2,12 +2,12 @@ import queue
 from typing import List
 from model.dto.file_model_dto import FileModelDTO
 
+POISON_PILL = FileModelDTO(job_name="_poison_pill_", name="_poison_pill_", priority=0)
+
 
 class FileQueue(queue.PriorityQueue):
     """A priority queue for files. Allows updating the priority of files
     already in the queue."""
-
-    POISON_PILL = FileModelDTO(job_name="_poison_pill_", name="_poison_pill_", priority=0)
 
     def __init__(self):
         """Create a new file queue."""
@@ -20,11 +20,15 @@ class FileQueue(queue.PriorityQueue):
 
     def is_poison_pill(entry: FileModelDTO) -> bool:
         """Returns True if the tested entry is a poison pill, False otherwise."""
-        return entry.name is not None and entry.name == FileQueue.POISON_PILL.name
+        return (
+            entry is not None
+            and entry.name is not None
+            and entry.name == POISON_PILL.name
+        )
 
     def poison_pill(self) -> None:
         """Put an entry to the queue which is by convention a poison pill."""
-        self.put_file(FileQueue.POISON_PILL)
+        self.put_file(POISON_PILL)
 
     def put_all(self, files: List[FileModelDTO]) -> None:
         """Put all files into the queue with the given priority.
@@ -44,15 +48,19 @@ class FileQueue(queue.PriorityQueue):
         """Put a file into the queue.
         :param file:
             The file to put into the queue"""
-        if file is None:
-            raise ValueError(
-                "Can't add None to this queue. Use .poison_pill() instead."
-            )
-        if file.name in self.entry_finder:
-            self.remove_file(file)
-        entry = [file.priority, file]
-        self.entry_finder[file.name] = entry
-        self.put(entry)
+        # poison pill has a priority treatment
+        if FileQueue.is_poison_pill(file):
+            self.put([0, file])
+        else:
+            if file is None:
+                raise ValueError(
+                    "Can't add None to this queue. Use .poison_pill() instead."
+                )
+            if file.name in self.entry_finder:
+                self.remove_file(file)
+            entry = [file.priority, file]
+            self.entry_finder[file.name] = entry
+            self.put(entry)
 
     def remove_file(self, file: FileModelDTO) -> None:
         """Remove a file from the queue.
@@ -70,6 +78,6 @@ class FileQueue(queue.PriorityQueue):
         _, file = self.get()
         while file is self.REMOVED:
             _, file = self.get()
-        if file is not None:
+        if file is not None and not FileQueue.is_poison_pill(file):
             del self.entry_finder[file.name]
         return file
